[Skip] improve performance enumerator.

+++ src/StructLinq.Tests/RefSkipTests.cs
@@ -20,9 +20,9 @@ namespace StructLinq.Tests
         [InlineData(0)]
         [InlineData(5)]
         [InlineData(10)]
-        public void ShouldBeTheSameAsSystem(uint skipCount)
+        public void ShouldBeTheSameAsSystem(int skipCount)
         {
-            var expected = Enumerable.Range(0, 7).ToArray().Skip((int)skipCount).ToArray();
+            var expected = Enumerable.Range(0, 7).ToArray().Skip(skipCount).ToArray();
             var value = Enumerable.Range(0, 7).ToArray().ToRefStructEnumerable().Skip(skipCount).ToArray();
 
             Assert.Equal(expected, value);
