+++ src/StructLinq/Skip/StructCollection.Skip.cs
@@ -7,40 +7,40 @@ namespace StructLinq
     public static partial class StructEnumerable
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static TStructCollection Skip<T,TStructCollection, TEnumerator>(this TStructCollection collection, uint count, Func<TStructCollection, IStructCollection<T, TEnumerator>> _)
+        public static TStructCollection Skip<T,TStructCollection, TEnumerator>(this TStructCollection collection, int count, Func<TStructCollection, IStructCollection<T, TEnumerator>> _)
             where TEnumerator : struct, IStructEnumerator<T>
             where TStructCollection : struct, IStructCollection<T, TEnumerator>
         {
             var newCollection = collection;
-            newCollection.Slice(count, null);
+            newCollection.Slice((uint)count, null);
             return newCollection;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static IStructCollection<T, TEnumerator> Skip<T, TEnumerator>(this IStructCollection<T, TEnumerator> collection, uint count)
+        public static IStructCollection<T, TEnumerator> Skip<T, TEnumerator>(this IStructCollection<T, TEnumerator> collection, int count)
             where TEnumerator : struct, IStructEnumerator<T>
         {
             var newCollection = (IStructCollection<T, TEnumerator>)collection.Clone();
-            newCollection.Slice(count, null);
+            newCollection.Slice((uint)count, null);
             return newCollection;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static TStructCollection Skip<T,TStructCollection, TEnumerator>(this TStructCollection collection, uint count, Func<TStructCollection, IRefStructCollection<T, TEnumerator>> _)
+        public static TStructCollection Skip<T,TStructCollection, TEnumerator>(this TStructCollection collection, int count, Func<TStructCollection, IRefStructCollection<T, TEnumerator>> _)
             where TEnumerator : struct, IRefStructEnumerator<T>
             where TStructCollection : struct, IRefStructCollection<T, TEnumerator>
         {
             var newCollection = collection;
-            newCollection.Slice(count, null);
+            newCollection.Slice((uint)count, null);
             return newCollection;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static IRefStructCollection<T, TEnumerator> Skip<T, TEnumerator>(this IRefStructCollection<T, TEnumerator> collection, uint count)
+        public static IRefStructCollection<T, TEnumerator> Skip<T, TEnumerator>(this IRefStructCollection<T, TEnumerator> collection, int count)
             where TEnumerator : struct, IRefStructEnumerator<T>
         {
             var newCollection = (IRefStructCollection<T, TEnumerator>)collection.Clone();
-            newCollection.Slice(count, null);
+            newCollection.Slice((uint)count, null);
             return newCollection;
         }
     }
